Implemented the setup_activity into the dashboard activity. Now the authenticated user can upload a picture of himself and store his name. These appear even when you browse across different views and come back to the dashboard : the image and name are therefore correctly retrieved.
TODO : implement now the featuring fragment and the showing of the old posts now that we have the picture of the user available

TODO : the image appearing in the dashboard depends on the size of the cropped image yet I don't want the circular image to change size depending on the crop size. Change this.