Implemented GlideApp and now the code to add an image from the storage in firebase to a view became simpler. When you go back to the dashboard the correct profile picture appears as wished. There are still problems related to the image :
1. The image is shown the actual size at which it is cropped. When the circle becomes smaller, then the profile picture becomes even more cropped.
2. The profile picture does not appear in the featuring section along with the username.

Could be an idea next to polish the dashboard, implement a good interface for the chat, add google sign in (taking into account the fact that it should be implemented with the authentication and the database), and add a search bar which will search for posts that have a relevant hashtag. But to do this, then I should allow the user to associate hashtags to the post.